refactor: move KeyList to kv module
