+++ src/commands/kv/key/mod.rs
@@ -1,11 +1,9 @@
 mod delete;
 mod get;
-mod key_list;
 mod list;
 mod put;
 
 pub use delete::delete;
 pub use get::get;
-pub use key_list::KeyList;
 pub use list::list;
 pub use put::{put, KVMetaData};
