+++ src/kv/mod.rs
@@ -1,2 +1,3 @@
 pub mod bulk;
+pub mod key;
 pub mod namespace;
